Finalizado CRUD de usuários

+++ resources/views/user/index.blade.php
@@ -22,6 +22,8 @@
                             <th>Código</th>
                             <th>Nome</th>
                             <th>Perfil</th>
+                            <th>Criado em</th>
+                            <th>Atualizado em</th>
                             <th>Ações</th>
                         </tr>
                     </thead>
@@ -31,14 +33,28 @@
                                 <td>{{ sprintf('%03d', $user->id) }}</td>
                                 <td>{{ $user->name }}</td>
                                 <td>{{ $user->profile_description }}</td>
+                                <td>{{ $user->created }}</td>
+                                <td>{{ $user->updated }}</td>
                                 <td>
-                                    <a href="{{ route('user.edit', $user->id) }}" alt="Editar usuário" title="Editar usuário" class="btn btn-primary btn-sm"><i class="fa fa-pencil"></i></a>
                                     
                                     <a href="{{ route('user.show', $user->id) }}" alt="Visualizar informações do usuário" title="Visualizar informações do usuário" class="btn btn-default btn-sm"><i class="fa fa-eye"></i></a>                                    
                                     
-                                    {{ Form::open(['method' => 'DELETE', 'route' => ['user.destroy', $user->id], 'style' => 'display: inline', 'onsubmit' => 'return confirmDelete()']) }}
-                                        {{ Form::button('<i class="fa fa-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-sm', 'alt' => 'Deletar usuário', 'title' => 'Deletar usuário']) }}
-                                    {{ Form::close() }}
+                                    @if (Auth::user()->isSuperAdmin())
+                                        <a href="{{ route('user.edit', $user->id) }}" alt="Editar usuário" title="Editar usuário" class="btn btn-primary btn-sm"><i class="fa fa-pencil"></i></a>
+
+                                        {{ Form::open(['method' => 'DELETE', 'route' => ['user.destroy', $user->id], 'style' => 'display: inline', 'onsubmit' => 'return confirmDelete()']) }}
+                                            {{ Form::button('<i class="fa fa-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-sm', 'alt' => 'Deletar usuário', 'title' => 'Deletar usuário']) }}
+                                        {{ Form::close() }}
+                                    @else
+                                        @if ($user->profile != config('profiles.superadmin'))
+                                            <a href="{{ route('user.edit', $user->id) }}" alt="Editar usuário" title="Editar usuário" class="btn btn-primary btn-sm"><i class="fa fa-pencil"></i></a>
+
+                                            {{ Form::open(['method' => 'DELETE', 'route' => ['user.destroy', $user->id], 'style' => 'display: inline', 'onsubmit' => 'return confirmDelete()']) }}
+                                                {{ Form::button('<i class="fa fa-trash"></i>', ['type' => 'submit', 'class' => 'btn btn-danger btn-sm', 'alt' => 'Deletar usuário', 'title' => 'Deletar usuário']) }}
+                                            {{ Form::close() }}
+                                        @endif
+                                    @endif
+
                                 </td>
                             </tr>
                         @endforeach
